Connect admin service to product service and add new swagger for customer service

+++ CustomerService/app/Services/OrderService.php
@@ -8,6 +8,7 @@ use App\Http\Resources\OrderCollection;
 use App\Http\Resources\OrderResource;
 use App\Repositories\IOrderRepository;
 use Illuminate\Auth\Access\AuthorizationException;
+use Illuminate\Support\Facades\Http;
 
 class OrderService implements IOrderService
 {
@@ -27,7 +28,17 @@ class OrderService implements IOrderService
 
     public function storeOrder(int $userId, StoreOrderRequest $request)
     {
-        $order = $this->orderRepository->storeOrder($userId, $request);
+        $response = Http::post(env('PRODUCT_SERVICE_URL') . '/products/get', [
+            'product_ids' => $request['product_ids']
+        ]);
+
+        if ($response->status() !== 200) {
+            throw new NotFoundException("Nie znaleziono produktów.");
+        }
+
+        $products = $response->json();
+
+        $order = $this->orderRepository->storeOrder($userId, $request, $products);
 
         return new OrderResource($order);
     }
@@ -38,6 +49,7 @@ class OrderService implements IOrderService
 
         if (!isset($order)) throw new NotFoundException();
         if ($order['user_id'] !== $userId) throw new AuthorizationException();
+        if ($order['status'] !== 'pending') throw new AuthorizationException();
 
         $this->orderRepository->destroyOrder($userId, $orderId);
     }
